# docs/swagger.yaml
basePath: /
definitions:
...
  keerja-backend_internal_domain_master.CreateJobTitleRequest:
    properties:
      name:
        maxLength: 200
        minLength: 2
        type: string
      popularity_score:
        maximum: 100
        minimum: 0
        type: number
      recommended_category_id:
        type: integer
    required:
    - name
    type: object
  keerja-backend_internal_domain_master.UpdateJobTitleRequest:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 200
        minLength: 2
        type: string
      popularity_score:
        maximum: 100
        minimum: 0
        type: number
      recommended_category_id:
        type: integer
    type: object
...
  keerja-backend_internal_utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@keerja.com
    name: API Support
  description: Job platform backend API with authentication, job management, applications,
    companies, and push notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Keerja Backend API
  version: "1.0"
paths:
...
  /api/v1/admin/master/job-titles:
    post:
      consumes:
      - application/json
      description: Create a new job title with category recommendation
      parameters:
      - description: Job title details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keerja-backend_internal_domain_master.CreateJobTitleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new job title (admin only)
      tags:
      - Admin - Master Data
  /api/v1/admin/master/job-titles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a job title
      parameters:
      - description: Job Title ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a job title (admin only)
      tags:
      - Admin - Master Data
    get:
      consumes:
      - application/json
      description: Retrieve a single job title by ID
      parameters:
      - description: Job Title ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get a job title by ID (admin only)
      tags:
      - Admin - Master Data
    put:
      consumes:
      - application/json
      description: Update an existing job title
      parameters:
      - description: Job Title ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated job title details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keerja-backend_internal_domain_master.UpdateJobTitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a job title (admin only)
      tags:
      - Admin - Master Data
...
  /api/v1/master/job-options:
    get:
      consumes:
      - application/json
      description: Get job types, work policies, education levels, experience levels,
        and gender preferences
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get all job posting options (DEPRECATED - use /job-types instead)
      tags:
      - Master Data
  /api/v1/master/job-posting-form-options:
    get:
      consumes:
      - application/json
      description: Returns job categories, subcategories, job types, work policies,
        education levels, experience levels, gender preferences, salary and age defaults
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get job posting form fields/options for mobile
      tags:
      - Master Data
  /api/v1/master/job-titles:
    get:
      consumes:
      - application/json
      description: Search job titles with fuzzy matching and category recommendations
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: 'Results limit (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get job titles with smart search
      tags:
      - Master Data
  /api/v1/master/job-types-options:
    get:
      consumes:
      - application/json
      description: Get job types, work policies, company addresses, and salary ranges
        for job posting
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get job types and work policies for mobile
      tags:
      - Master Data
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"